+++ lowlevel/src/utility/PIDFP.h
@@ -17,7 +17,7 @@
 class PID_FP {
   uint32_t kp, ki, kd;
   int32_t min, max;
-  int32_t freq;
+  uint32_t freq;
   int64_t integral_max, integral_min;
 
   int32_t derivative_max, derivative_min;
@@ -31,13 +31,13 @@ class PID_FP {
 
   void reset();
 
-  void set_coefficients(float new_kp, float new_ki, float new_kd, uint32_t new_freq);
+  void set_coefficients(double new_kp, double new_ki, double new_kd, uint32_t new_freq);
 
-  void set_kp(uint32_t new_kp);
+  void set_kp(double new_kp);
 
-  void set_ki(uint32_t new_ki, uint32_t new_freq);
+  void set_ki(double new_ki, uint32_t new_freq);
 
-  void set_kd(uint32_t new_kd, uint32_t new_freq);
+  void set_kd(double new_kd, uint32_t new_freq);
 
   void set_output_limit(int32_t new_limits);
 
@@ -45,7 +45,7 @@ class PID_FP {
 
   void set_anti_windup(int32_t new_limit);
 
-  void get_coefficients(float *ret_kp, float *ret_ki, float *ret_kd);
+  void get_coefficients(double *ret_kp, double *ret_ki, double *ret_kd);
 
   int16_t compute(int32_t input, int32_t setpoint);
